% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecode-package.R
\docType{package}
\name{ecode-package}
\alias{ecode}
\alias{ecode-package}
\title{ecode: Ordinary Differential Equation Systems in Ecology}
\description{
Provides general framework to investigate population and community dynamics through dyanmical system modelling with integration of a wide array of graphical, analytical, and numerical techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HaoranPopEvo/ecode}
  \item Report bugs at \url{https://github.com/HaoranPopEvo/ecode/issues}
}

}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

Authors:
\itemize{
  \item Haoran Wu \email{haoran.wu@wolfson.ox.ac.uk} (\href{https://orcid.org/0000-0001-9449-6112}{ORCID})
}

}
\keyword{internal}
