% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype.R
\name{pdist}
\alias{pdist}
\title{Distance between Phase Points}
\usage{
pdist(x, y)
}
\arguments{
\item{x}{an object of \code{"pp"} class representing a phase point.}

\item{y}{an object of \code{"pp"} class representing another phase point.}
}
\value{
A scalar.
}
\description{
Computes and returns the distance between two phase points.
}
\examples{
a <- pp(list(x = 1, y = 1))
b <- pp(list(x = 3, y = 4))
pdist(a, b)
}
