% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype.R
\name{pp}
\alias{pp}
\title{Create a Phase Point}
\usage{
pp(value)
}
\arguments{
\item{value}{A list of several elements, each giving a value on one dimension.}
}
\value{
An object of class "\code{pp}" describing a phase point.
}
\description{
Creates an object of class "\code{pp}" representing a phase point in the phase
space.
}
\examples{
pp(list(x = 1, y = 1))
}
