% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype.R
\name{print.pp}
\alias{print.pp}
\title{Print Brief Details of a Phase Point}
\usage{
\method{print}{pp}(x, ...)
}
\arguments{
\item{x}{Object of class "\code{pp}".}

\item{...}{Ignored.}
}
\value{
No value
}
\description{
Prints a very brief description of a phase point.
}
\examples{
pp(list(x = 1, y = 1))
}
