% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype.R
\name{*.pp}
\alias{*.pp}
\title{Multiply Operator for Phase Points}
\usage{
\method{*}{pp}(x, y)
}
\arguments{
\item{x}{an object of \code{"pp"} class representing a phase point.}

\item{y}{an object of \code{"pp"} class representing another phase point.}
}
\value{
an object of \code{"pp"} class representing the calculated result.
}
\description{
The multiply operator used to perform arithmetic on phase points.
}
\examples{
pp(list(x = 1, y = 1)) * pp(list(x = 3, y = 4))
}
