% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecode-package.R
\docType{package}
\name{ecode-package}
\alias{ecode}
\alias{ecode-package}
\title{ecode: Ordinary Differential Equation Systems in Ecology}
\description{
A framework to simulate ecosystem dynamics through ordinary differential equations (ODEs). You create an ODE model, tells 'ecode' to explore its behaviour, and perform numerical simulations on the model. 'ecode' also allows you to fit model parameters by machine learning algorithms. Potential users include researchers who are interested in the dynamics of ecological community and biogeochemical cycles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HaoranPopEvo/ecode}
  \item Report bugs at \url{https://github.com/HaoranPopEvo/ecode/issues}
}

}
\author{
\strong{Maintainer}: Haoran Wu \email{haoran.wu@wolfson.ox.ac.uk} (\href{https://orcid.org/0000-0001-9449-6112}{ORCID})

Authors:
\itemize{
  \item Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})
}

}
\keyword{internal}
